Refactor launcher into focused modules
